+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - `katex`: `cssFile`, `placeholder` and `fontsFolder` option. The plugin downloads the CSS code.
 - `icons` plugin: Added css.gg and radix-ui.com/icons [#736]
 - `feed` plugin: Allow to pass an array of options or a function that return an array of options.
+- `sitemap` plugin: New option `stylesheet`.
 
 ### Changed
 - `Temporal` API is enabled by default.